requirements.txt
# Core AI/ML Framework
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# Database Drivers
psycopg2-binary>=2.9.0
neo4j>=5.0.0
pymilvus>=2.3.0

# Environment & HTTP
python-dotenv>=1.0.0
requests>=2.31.0

# LLM Providers
google-generativeai>=0.3.0
google-api-python-client>=2.0.0
openai>=1.0.0
anthropic>=0.7.0

# Utilities
python-multipart>=0.0.6
jinja2>=3.1.0
markdown>=3.5.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
